36-spectra-iq-spectrum-awareness/requirements.txt
streamlit>=1.30
plotly>=5.20
scipy>=1.11
numpy>=1.24
pandas>=2.0
Pillow>=10.0
openai>=1.40
python-dotenv>=1.0
